split instance resurrection into a separate package
All this has to do is watch for delete events and orchestrate the node.
Trivial to break out.

Also removed the suicidal node deal. I'd rather do it in a way that literally
pkills the node and relies on the keys expiring. A similar mechanism may be
useful for testing evacuation, but I want to get to the point where I can 1.
have this guy stop responding to the instances, and 2. not have such hacky
code.